Don't pollute the Model and Resource namespaces
All the methods are still available via the #i18n
proxy accessor available in both DataMapper::Model
and DataMapper::Resource.

Arguably, these methods are mostly useful for code
that handles UI for resource translations. I think
that we are not violating the Law of Demeter here,
but instead shifting responsiblity to where it
belongs, while at the same time not exposing too
many methods for the common case.