Process related arguments in one go
The `model`, `options` and `&block` variables only
make sense when used together. Having the handling
of these parameters separated into two method calls
(#initialize and #localize) felt weird.

Also, it can be argued that after this refactoring,
the Localizer class is ready to actually localize
a model, right after #initialize was called.

The actual work of localizing a model:

    setup_proxy_accessor_api
    relate_translation_model
    generate_accessor_aliases
    generate_property_readers

can then be done inside a single call to #localize