Add support for namespaced translation models
The namespace will be determined by the following
rules (FQN stands short for Fully Qualified Name)

1) If :namespace is passed to #translatable, the
   value must be a const and will be used:

   module Translations; end
   My::Story.translatable :namespace => Translations do
     # FQN: Translations::StoryTranslation
   end

2) If no :namespace is passed to #translatable and
   DataMapper::I18n.translation_model_namespace is
   nil, the translated model's namespace will be
   used (i.e. :namespace => self)

   My::Story.translatable do
     # FQN: My::StoryTranslation
   end

3) If no :namespace is passed to #translatable and
   DataMapper::I18n.translation_model_namespace is
   not nil, this will be used:

   DataMapper::I18n.translation_model_namespace =
     ContentTranslations
   My::Story.translatable do
     # FQN: ContentTranslations::StoryTranslation
   end

Furthermore, the :model option for #translatable
now always demodulizes the passed String, therefore
passing fully qualified names makes no sense, as
those will be stripped and the :namespace option
will be used to determine the namespace of the
generated model. Of course it's still possible to
change the generated model name by passing the
desired name to the :model option.