Tentative support for translating models with a CPK
TODO: This still needs specs! The way the current
specsuite is organized doesn't make it particularly
easy to come up with proper specs for that case
right now, so I'm gonna add them once i find time
to rewrite the specs. Since I'll be needing this
functionality in a different project soonish, I'm
going to use that for testing, and fix any bugs
before writing the dedicated specs, should that be
necessary.